requirements.txt
discord.py
motor
lru-dict
python-dotenv